Various enhancements - (WIP) Remove SQLiteCpp - Display live voice availability in single mode race results - Add il2cppFnMapPath option #16 - Change injection method #16 - Remove deprecated behavior

+++ src/rich_presence.hpp
@@ -9,9 +9,9 @@ using namespace std;
 
 namespace
 {
-	extern unsigned long GetTextIdByName(const wstring& name);
+	extern uint64_t GetTextIdByName(const wstring& name);
 
-	extern Il2CppString* localizeextension_text_hook(int id);
+	extern Il2CppString* localizeextension_text_hook(uint64_t id);
 
 	string GetViewName(string viewId)
 	{
